requirements-production.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
Pillow>=10.0.0
numpy<2.0.0,>=1.24.0
boto3>=1.26.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
structlog>=23.0.0
tenacity>=8.0.0